+++ app/src/main/java/com/swallow/cracker/data/model/RedditChildrenResponse.kt
@@ -1,8 +1,10 @@
 package com.swallow.cracker.data.model
 
+import com.squareup.moshi.Json
 import com.squareup.moshi.JsonClass
 
 @JsonClass(generateAdapter = true)
 data class RedditChildrenResponse(
+    @Json(name = "data")
     val data: RedditNewsDataResponse
 )
