+++ app/src/main/java/com/swallow/cracker/ui/activitys/MainActivity.kt
@@ -3,4 +3,4 @@ package com.swallow.cracker.ui.activitys
 import androidx.appcompat.app.AppCompatActivity
 import com.swallow.cracker.R
 
-class MainActivity : AppCompatActivity(R.layout.activity_main) {}
+class MainActivity : AppCompatActivity(R.layout.activity_main)
