+++ app/src/main/java/com/swallow/cracker/ui/adapters/OnBoardingAdapter.kt
@@ -7,7 +7,7 @@ import com.swallow.cracker.databinding.OnboardingItemBinding
 import com.swallow.cracker.ui.adapters.viewholders.OnBoardingViewHolder
 import com.swallow.cracker.ui.model.OnBoardingUI
 
-class OnBoardingAdapter() : RecyclerView.Adapter<OnBoardingViewHolder>() {
+class OnBoardingAdapter : RecyclerView.Adapter<OnBoardingViewHolder>() {
 
     private var items: List<OnBoardingUI> = listOf()
