+++ app/src/main/java/com/swallow/cracker/ui/adapters/delegates/ComplexDelegateAdapterClick.kt
@@ -1,0 +1,10 @@
+package com.swallow.cracker.ui.adapters.delegates
+
+import com.swallow.cracker.ui.model.RedditItems
+
+interface ComplexDelegateAdapterClick {
+    fun onVoteClick(position: Int, likes: Boolean)
+    fun onSavedClick(category: String?, id: String, position: Int?, saved: Boolean)
+    fun navigateTo(item: RedditItems)
+    fun shared(url: String)
+}
